fix(overview): remove incorrect ifconfig use
Remove the last of the incorrect usage of the ifconfig directive.
Replace these cumbersome tables with list-tables.

This is a little goofy due to old formatting. To keep in line with what
was previously here (and to prevent automatic numbering of these
tables), we must place the header in the first row and pad accordingly.

Signed-off-by: Randolph Sapp <rs@ti.com>