fix(gen): purge gen configs and sections
Remove the GEN platform configs for linux and android and all hidden
content in the current pages.

This is a meta-platform that was supposed to enable show a common
document for all arm32 platforms. Unfortunately this contained a lot of
hidden information that was not being exposed for standard releases and
is not being updated.

Signed-off-by: Randolph Sapp <rs@ti.com>