docs(readme): simplify the readme and example make
At this point we should drive users to the container properly. That
gives them access to almost all of the CI tools and ensures the output
they generate will match what we expect. Advanced users can still parse
the requirements.txt if they want.

Also, add an example make command using the container to the top of the
build guide since that is one of the more popular things I need to
remind people about.

Signed-off-by: Randolph Sapp <rs@ti.com>